requirements.txt
dacite
weasyprint
Jinja2
PyYAML
